+++ gaios/Services/SessionManager.swift
@@ -343,7 +343,6 @@ class SessionManager {
                 }
                 return Promise<Void>().asVoid()
             }.map(on: bgq) {
-                self.registry?.cache(session: self)
                 self.registry?.loadAsync(session: self)
             }.tapLogger()
     }
@@ -352,10 +351,7 @@ class SessionManager {
         let bgq = DispatchQueue.global(qos: .background)
         return Guarantee()
             .then(on: bgq) { self.loginUser(details: ["username": username, "password": password]) }
-            .get { _ in
-                self.registry?.cache(session: self)
-                self.registry?.loadAsync(session: self)
-            }
+            .get { _ in self.registry?.loadAsync(session: self) }
     }
 
     func loginWithHW(_ hwDevice: HWDevice) -> Promise<String> {