+++ CHANGELOG.md
@@ -12,9 +12,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Announcements alerts: improve navigation
+- Improved performance of asset registry usage
 
 ### Fixed
-
 
 ## [3.8.8] - 2022-09-23